--- src/common/models/job/index.ts
@@ -1,2 +1,0 @@
-export { default as UploadJob, Options as UploadJobOptions } from './upload-job'
-export { default as DownloadJob } from './download-job'
