+++ src/common/models/job/types.ts
@@ -13,3 +13,21 @@ export enum Status {
     Duplicated = "duplicated",
     Verifying = "verifying"
 }
+
+export interface LocalPath {
+    name: string,
+    path: string,
+    size?: number, // bytes
+    mtime?: number, // ms timestamp
+}
+
+export interface RemotePath {
+    bucket: string,
+    key: string,
+    size?: number, // bytes
+    mtime?: number, // ms timestamp
+}
+
+export function isLocalPath(p: LocalPath | RemotePath): p is LocalPath {
+    return p.hasOwnProperty("name");
+}
