+++ src/common/qiniu/index.ts
@@ -1,2 +1,3 @@
 export * from "./types";
 export {default as createQiniuClient} from "./create-client";
+export * from "./etag";
