--- src/main/transfer-managers/remote-walker.ts
@@ -1,19 +1,0 @@
-// import {Adapter} from "kodo-s3-adapter-sdk/dist/adapter";
-//
-// function CreateRemoteWalker(client: Adapter) {
-//     async function _walkPage() {
-//
-//     }
-//
-//     async function _walk() {
-//
-//     }
-// }
-//
-// export default class RemoteWalker {
-//     constructor(
-//         private readonly client: Adapter,
-//
-//     ) {
-//     }
-// }
