+++ src/renderer/components/services/auto-upgrade.js
@@ -4,7 +4,7 @@ import path from 'path'
 import request from 'request'
 import downloadsFolder from 'downloads-folder'
 
-import * as util from '@common/models/job/utils'
+import {getEtag} from '@common/qiniu'
 import webModule from '@/app-module/web'
 
 import { upgrade } from '@/customize'
@@ -82,7 +82,7 @@ webModule.factory(AUTO_UPGRADE_SVS_FACTORY_NAME, [
       this.check = function (expected, callback) {
         //crc
         console.log("etag check");
-        return util.getEtag(to + ".download", function(actual) {
+        return getEtag(to + ".download", function(actual) {
           if (expected !== `"${actual}"`) {
             callback(new Error(`Etag check failed, expected: ${expected}, actual: "${actual}"`));
           } else {
